pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "fastchat-mcp"
version = "1.1.2"
authors = [{ name = "Raúl Beltrán Gómez", email = "rb58853@gmail.com" }]
description = "Python client, based on fastmcp, for connecting to MCP servers through multiple protocols, specifically designed to work with integrated language models."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICENSE*"]
dependencies = [
  "openai>=1.68.2",
  "mcp[cli]",
  "mcp-oauth>=0.0.5",
  "fastapi",
  "websockets",
  "requests",
  "fastauth-api",
]

[project.urls]
Homepage = "https://github.com/rb58853/fastchat-mcp"
Issues = "https://github.com/rb58853/fastchat-mcp/issues"
